refactor: move modal form in separate file, add inteface
