+++ src/interfaces/modalForm.ts
@@ -1,0 +1,3 @@
+export interface IModalForm {
+  setModalContent: (a: boolean) => void;
+}
