+++ src/sass/_variables.scss
@@ -6,7 +6,7 @@ $color-white: #ffffff;
 $color-dark-grey: #222523;
 $color-green: #13921e;
 $color-light-grey: #d9e5e1;
-
+$color-red: #ff0000;
 
 $color-black: #000;
 $color-white-grey:#f5f5f5;
